fix: try fileout

+++ src/post.js
@@ -1,6 +1,8 @@
 const pcapParser = require("./pcap_parser");
-const { sleepSync } = require('./sleep');
 const { exec } = require('child_process');
+const fs = require("fs")
+
+const { sleepSync } = require('./sleep');
 
 const filePcap = '/tmp/dns.pcap';
 
@@ -8,22 +10,29 @@ const supressOutput = process.env.SUPRESS_DNS_AUDIT_OUTPUT || false;
 
 const post = () => {
 
-    exec('sudo pkill tcpdump', (err, stdout, stderr) => {
-        if (err) {
-            console.log(err);
-            return;
+    // Only clean up if the file exists
+    if (fs.existsSync(filePcap)) {
+
+        exec('sudo pkill tcpdump', (err, stdout, stderr) => {
+            if (err) {
+                console.log(err);
+                return;
+            }
+            console.log(stdout);
+            console.log(stderr);
+        });
+
+        // Let tcpdump finish
+        sleepSync(5000);
+
+        // Convert PCAP to JSON
+        if (!supressOutput) {
+            const packets = pcapParser.parsePcapFile(filePcap);
+            console.log(packets);
         }
-        console.log(stdout);
-        console.log(stderr);
-    });
 
-    // Let tcpdump finish
-    sleepSync(5000);
-
-    // Convert PCAP to JSON
-    if (!supressOutput) {
-        const packets = pcapParser.parsePcapFile(filePcap);
-        console.log(packets);
+        // Delete file to avoid multiple runs
+        fs.unlinkSync(filePcap);
     }
 
 }
